# Cache digit width for tabular numbers