Fixes #470.
The max digit width is now calculated when the font is loaded, even if that font is never used for tabular numbers.