@CryZe 2021-12-11
Seems to have been a decent performance improvement:
```
Scene Management (Default)
Before:
                        time:   [7.3866 us 7.4272 us 7.4747 us]
After:
                        time:   [4.3817 us 4.4423 us 4.4991 us]

Scene Management (Subsplits Layout)
Before:
                        time:   [26.779 us 26.782 us 26.786 us]
After:
                        time:   [20.934 us 21.212 us 21.506 us]
```